Update _chips.scss
This is a minor but suitable change which updates the width of input field (for adding new chips) from fixed 120px to fit-content (wherever supported), with same !important weight. Also, the min-width and max-width properties are set to limit its size between 100px and 200px. With this change, the input placeholder (if limited to under 200px) will not get cropped, and it will also have a better fit into the input field, thereby offering it more flexibility to adapt the content.